refactor(pages): restructure page components for SSR support
- remove individual page components (HomePage, ProductDetailPage, NotFoundPage)
- add new page structure with server-side rendering support
- implement HomePage with renderHomePageForServer and getServerSideProps
- add dynamic product page with [id].js structure
- add dedicated error and loading pages for better UX
