--- packages/vanilla/src/pages/HomePage.js
@@ -1,100 +1,0 @@
-import { ProductList, SearchBar } from "../components";
-import { productStore } from "../stores";
-import { router, withLifecycle } from "../router";
-import { loadProducts, loadProductsAndCategories } from "../services";
-import { PageWrapper } from "./PageWrapper.js";
-import { filterProducts, getUniqueCategories, paginateProducts, items } from "../mocks/shared.js";
-import { hasHomePageData } from "../utils";
-
-// 실제 렌더링 로직 - 서버와 클라이언트에서 공통 사용
-const renderHomePage = (productState, query) => {
-  const { search: searchQuery, limit, sort, category1, category2 } = query;
-  const { products, loading, error, totalCount, categories } = productState;
-  const category = { category1, category2 };
-  const hasMore = products.length < totalCount;
-
-  return PageWrapper({
-    headerLeft: `
-      <h1 class="text-xl font-bold text-gray-900">
-        <a href="/" data-link>쇼핑몰</a>
-      </h1>
-    `.trim(),
-    children: `
-      <!-- 검색 및 필터 -->
-      ${SearchBar({ searchQuery, limit, sort, category, categories })}
-      
-      <!-- 상품 목록 -->
-      <div class="mb-6">
-        ${ProductList({
-          products,
-          loading,
-          error,
-          totalCount,
-          hasMore,
-        })}
-      </div>
-    `.trim(),
-  });
-};
-
-// 서버 전용 렌더 함수
-export const renderHomePageForServer = (initialData) => {
-  const productState = {
-    products: initialData.products,
-    loading: initialData.loading,
-    error: initialData.error,
-    totalCount: initialData.totalCount,
-    categories: initialData.categories,
-  };
-
-  return renderHomePage(productState, initialData.query);
-};
-
-// 클라이언트용 컴포넌트 (기존과 동일한 구조)
-export const HomePage = withLifecycle(
-  {
-    onMount: () => {
-      // SSR 데이터가 이미 있다면 로딩하지 않음
-      if (!hasHomePageData()) {
-        loadProductsAndCategories();
-      }
-    },
-    watches: [
-      () => {
-        const { search, limit, sort, category1, category2 } = router.query;
-        return [search, limit, sort, category1, category2];
-      },
-      () => loadProducts(true),
-    ],
-  },
-  () => {
-    const productState = productStore.getState();
-    const query = router.query;
-    return renderHomePage(productState, query);
-  },
-);
-
-export const getServerSideProps = async (context) => {
-  const query = context.query || {};
-  const page = Number(query.current ?? query.page ?? 1);
-  const limit = Number(query.limit ?? 20);
-  const sort = query.sort || "price_asc";
-  const normalizedQuery = { ...query, current: String(page), limit: String(limit), sort };
-
-  const filtered = filterProducts(items, normalizedQuery);
-  const pageProducts = paginateProducts(filtered, page, limit);
-  const categories = getUniqueCategories();
-  const totalCount = filtered.length;
-
-  return {
-    initialData: {
-      products: pageProducts,
-      categories,
-      totalCount,
-      loading: false,
-      error: null,
-      status: "done",
-      query: normalizedQuery,
-    },
-  };
-};
