+++ packages/vanilla/src/pages/not-found.js
