+++ packages/vanilla/src/pages/product/loading.js
@@ -1,0 +1,8 @@
+export const loadingContent = `
+  <div class="min-h-screen bg-gray-50 flex items-center justify-center">
+    <div class="text-center">
+      <div class="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-600 mx-auto mb-4"></div>
+      <p class="text-gray-600">상품 정보를 불러오는 중...</p>
+    </div>
+  </div>
+`;
